Fix PHP-FPM docroot setting for Apache 2.4
When using Apache 2.4, the document root is passed to the FPM process when connecting to the socket itself, rather than with each file request.  Without passing this value properly, apps with a docroot that isn't the project root (many frameworks fall under this category) will fail to load any files.