Add support for persistent Memcached sessions
The setting used to enable a persistent connection to Memcached for PHP session use is `session.save_path = "PERSISTENT={key} {host}:11211"` - the quotes, here, are essential to make the INI interpreter see anything after and including the embedded `=`. Adding the quotes directly in the template does no harm to other variants of this setting, and attempts to add them in the boxfile instead have met with .... interesting results. So, this change will make these kinds of configurations possible.